eir: P4.1 runtime shape tracking, dual bookkeeping + 64-bit header (joint gc-P1 layout)
Shapes-plan P4.1: the runtime grows a global interned type-aware shape
tree (runtime/ejs-shapes.{h,c}) mirroring maam's abstraction — a shape is
a (parent, name, repr) transition edge, repr in {f64, boxed} and part of
class identity.  Ordinary objects (_ejs_Object_specops) are born with the
root shape and keep their index maintained on insert/delete/type-flip;
the property map REMAINS the store — pure dual bookkeeping, zero behavior
change.  Deletes, non-default attributes, accessors, symbol/index keys,
and cap overflow drop the object to dictionary mode (shape 0) one-way,
with counted reasons.

Header: GCObjectHeader widens u32 -> u64 as the ONE joint gc-P1/shapes
layout (ejs-types.h documents the split: low 32 unchanged, bits 32-55
shape index, bit 56 P4.2 mode bit, 57-63 reserved for gc).  EJSObject /
EJSPrimString / EJSPrimSymbol sizes are unchanged (padding absorbed);
EJSClosureEnv grows 8 bytes (runtime-internal only).  lib/types.ts
mirrors the layout in this same commit, header as two i32 halves so
P4.3's has_shape can load the shape half directly; EJSModule field
indices and export-slot offsets are unchanged.

Perf: the per-parent transition memo is inlined into the insert hook
(a memo-hit name was vetted when the memo's shape was interned, so the
fast path is one ejsval compare); the hash cache keeps name hashes in
its entries to reject probe collisions without content compares.
Property-insert micro-overhead: 2.1% (mean of 5 interleaved runs, 300k
objects x 8 fresh atom-keyed inserts, worst case) vs the <5% gate; with
EJS_GC_DISABLE the on/off delta is zero.

Knobs: EJS_SHAPES=off (kill switch), EJS_SHAPES_CENSUS=1 (exit census:
shapes interned, transition/fast-hit counts, repr flips, migration
reasons, per-shape death counts), EJS_SHAPE_CAP (field cap, default 64).

Gates: matrix green (test-eir, lowtier, stages 0-3); the full stage1
suite is green under EJS_SHAPES=off via the new standing buck lane
//:test-stage1-shapes-off (buck-test-stage.sh grew a TEST_ENV arg);
census recorded in docs/shapes-plan.md (P4.1 ticked, numbers inline).

+++ runtime/ejs-types.h
@@ -27,7 +27,21 @@ typedef double jsdouble;
 
 typedef uint16_t jschar;
 
-typedef uint32_t GCObjectHeader;
+// The object header, widened to 64 bits as the joint gc-plan P1 /
+// shapes-plan P4.1 layout (one layout, written once — see
+// docs/gc-plan.md "Object header, forwarding, and shapes" and
+// docs/shapes-plan.md "Object layout, in two steps"):
+//
+//   bits  0-31  the pre-existing 32-bit header: EJSScanType in the low
+//               bits, user flags at EJS_GC_USER_FLAGS_SHIFT (unchanged)
+//   bits 32-55  shape index (0 = dictionary mode / untracked)
+//   bit  56     shaped-storage mode bit (reserved for shapes P4.2)
+//   bits 57-63  reserved for the GC (forwarding/age/mark/card, gc-P1)
+//
+// EJSObject absorbs the widening into what was padding (sizeof
+// unchanged); EJSPrimString/EJSPrimSymbol keep their sizes; EJSClosureEnv
+// grows by 8.  lib/types.ts mirrors this in the same commit.
+typedef uint64_t GCObjectHeader;
 
 #if defined(__GNUC__) && (__GNUC__ > 2)
 # define EJS_LIKELY(x)   (__builtin_expect((x), 1))
